refactor: remove centralized SpacePermission and SpaceEntityPermission interface and create separate ones for Fuse and High5
