+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "hcloud-sdk",
-    "version": "0.0.92",
+    "version": "0.0.93",
     "description": "hcloud platform sdk",
     "main": "index.js",
     "types": "index.d.ts",
